1、优化Grafana中deleteDashboard参数 2、补全清除MonitorGroup时的流程 删除数据库->删除grafana dashboard->删除文件系统image->删除dockerImage 这里有一个坑, 清除动作的中间部分是无法回滚的，那么一旦某个动作失败造成数据库回滚，那么下次再进行删除时就会出现错误，此处仍需要修改

+++ kafka-eagle-web/src/main/java/org/smartloli/kafka/eagle/web/service/impl/MonitorGroupServiceImpl.java
@@ -204,19 +204,26 @@ public class MonitorGroupServiceImpl implements MonitorGroupService {
             return new ValidateResult(ValidateResult.ResultCode.FAILURE, "数据删除失败");
         logger.info("========数据删除完成========");
 
+        // 删除grafana dashboard
+        ValidateResult deleteResult = grafanaDashboardService.deleteDashboard(monitorGroupId);
+        if(deleteResult.getResultCode() != ValidateResult.ResultCode.SUCCESS)
+            throw new NormalException(deleteResult.getMes()) ;
+        logger.info("========dashboard删除完成========");
+
+        // 删除image文件
+        //String path = monitorGroup.getPath();
+        //if(StringUtils.isBlank(path) ||c
+        //        Files.deleteIfExists(Paths.get(monitorGroup.getPath())))
+        //    return new ValidateResult(ValidateResult.ResultCode.FAILURE, "删除镜像文件失败!");
+        // logger.info("========image文件删除完成========")
+
         String res = dockerRestService.deleteMonitorImage(monitorGroup.getImageId());
         logger.info(res);
         Map<String, String> resMes = DockerRestResolver.resolveResult(res);
         if(!"200".equals(resMes.get("root.code")))
-            return new ValidateResult(ValidateResult.ResultCode.FAILURE, "镜像删除失败");
+            throw new NormalException("远程镜像删除失败");
 
-        // 删除image文件
-        //String path = monitorGroup.getPath();
-        //if(StringUtils.isBlank(path) ||
-        //        Files.deleteIfExists(Paths.get(monitorGroup.getPath())))
-        //    return new ValidateResult(ValidateResult.ResultCode.FAILURE, "删除镜像文件失败!");
-
-        return new ValidateResult(ValidateResult.ResultCode.SUCCESS, "镜像删除成功");
+        return new ValidateResult(ValidateResult.ResultCode.SUCCESS, "image删除成功");
     }
 
     @Override
